Make IMessage abstract with default implementation
