+++ autobahn/src/main/java/io/crossbar/autobahn/wamp/interfaces/ISerializer.java
@@ -11,16 +11,44 @@
 
 package io.crossbar.autobahn.wamp.interfaces;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
+import java.io.IOException;
 import java.util.List;
 
 
-public interface ISerializer {
+public abstract class ISerializer {
 
-    byte[] serialize(List<Object> message);
+    private final ObjectMapper mMapper;
 
-    List<Object> unserialize(byte[] payload, boolean isBinary);
+    public ISerializer(JsonFactory factor) {
+        mMapper = new ObjectMapper(factor);
+    }
 
-    <T> T convertValue(Object fromValue, TypeReference toValueTypeRef);
+    public byte[] serialize(List<Object> message) {
+        try {
+            return mMapper.writeValueAsBytes(message);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    public List<Object> unserialize(byte[] payload, boolean isBinary) {
+        try {
+            return mMapper.readValue(payload, new TypeReference<List<Object>>() {});
+        } catch (IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    public <T> T convertValue(Object fromValue, TypeReference toValueTypeRef) {
+        // https://github.com/FasterXML/jackson-databind#tutorial-fancier-stuff-conversions
+        // ResultType result = mapper.convertValue(sourceObject, ResultType.class);
+        return mMapper.convertValue(fromValue, toValueTypeRef);
+    }
 }
