+++ autobahn/src/main/java/io/crossbar/autobahn/wamp/serializers/JSONSerializer.java
@@ -12,46 +12,13 @@
 package io.crossbar.autobahn.wamp.serializers;
 
 import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.io.IOException;
-import java.util.List;
 
 import io.crossbar.autobahn.wamp.interfaces.ISerializer;
 
-public class JSONSerializer implements ISerializer {
 
-    public final ObjectMapper mMapper;
+public class JSONSerializer extends ISerializer {
 
     public JSONSerializer() {
-        mMapper = new ObjectMapper(new JsonFactory());
-    }
-
-    @Override
-    public byte[] serialize(List<Object> message) {
-        try {
-            return mMapper.writeValueAsBytes(message);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    @Override
-    public List<Object> unserialize(byte[] payload, boolean isBinary) {
-        try {
-            return mMapper.readValue(payload, new TypeReference<List<Object>>() {});
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    @Override
-    public <T> T convertValue(Object fromValue, TypeReference toValueTypeRef) {
-        // https://github.com/FasterXML/jackson-databind#tutorial-fancier-stuff-conversions
-        return mMapper.convertValue(fromValue, toValueTypeRef);
+        super(new JsonFactory());
     }
 }
